Finished implementing Basic MySQL connection

+++ src/main/java/com/mndk/kvm2m/core/vmap/VMapDataPayload.java
@@ -6,7 +6,7 @@ import lombok.ToString;
 
 import java.util.HashMap;
 
-public class VMapDataPayload extends HashMap<Integer, VMapDataPayload.Record> {
+public class VMapDataPayload extends HashMap<Long, VMapDataPayload.Record> {
 
     @RequiredArgsConstructor
     @Getter
